отображение экрана отсутствия предложений
